… .env.local.example
… bun.lockb
… lib/utils.ts