## Replaced privy, wagmi, viem, alchemy and biconomy with thirdweb

### Same features as before with less code
- connect any external wallet or use email/sms/socials - replaces privy
- read contract state, wallet balance - replaces wagmi/viem
- dedicated RPC - replaces alchemy
- connect to smart account and send gasless transactions - replaces biconomy

### Bonus features that came for free from this migration
- sign in with passkey
- connect with 350+ external wallets
- connected UI button - wallet logo, balance, ens resolution
- connected UI modal  - send / receive funds, purchase eth or erc20 with fiat or crypto, chain switcher
- auto refresh of contract state

### Performance improvements

|  Metric  | BEFORE | AFTER |
| ------------- | ------------- | ------------- |
| First Load JS  | `878 kB`  | `295 kB` - 66.4% lighter |
| Lighthouse Score   | `54`  | `93` - 72% higher |
| Speed index   | `1.1s`  | `0.7s` - 36.3 % faster |
| Report | ![wagmi](https://github.com/joaquim-verges/simple-dapp/assets/3353417/f23de926-db9a-4d41-bef0-53afb4d68b59) | ![thirdweb](https://github.com/joaquim-verges/simple-dapp/assets/3353417/25974d38-bb36-41d4-90db-14ee879a2cb1) |

Scores obtained running lighthouse on incognito chrome, cold start (no local data).